Grid and topo creation
Now the model is capable to create a meander grid in .grd Delft3D format and a topography file in .dep format that matches the grid.
Still I need to work on making the process fully automated (at the moment the wldep routine for creating the topography is called in interactive way.
I have tested the tool only in boundary fitted grid creation mode (not when the grid spans the full floodplain). Moreover, the model needs some cleanup because at present is creating the grid twice with different resolution parameters (one for Matlab visualisation, and one for Delft3D export), but this does not seem reasonable.